# mental-health-frontend-min (2)/微信接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "微信接口",
    "description": "前端 API 接口文档转换为 OpenAPI 3.0 格式，用于导入 Apifox。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/activity/checkin": {
      "post": {
        "summary": "学生扫码签到",
        "deprecated": false,
        "description": "学生通过扫码签到\n代码位置: `pages/activity/my-activity.js` - `realCheckIn()`",
        "operationId": "checkInActivity",
        "tags": [
          "微信前端/活动签到",
          "活动签到"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activityId": {
                    "type": "string"
                  },
                  "qrCodeData": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/activity/signup": {
      "post": {
        "summary": "报名活动",
        "deprecated": false,
        "description": "学生报名活动\n代码位置: `pages/activity/detail.js` - `handleSignup()`",
        "operationId": "signupActivity",
        "tags": [
          "微信前端/活动报名与审核",
          "活动报名与审核"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activityId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/activity/cancel": {
      "post": {
        "summary": "取消报名活动",
        "deprecated": false,
        "description": "学生取消报名活动\n代码位置: `pages/activity/detail.js` - `handleSignup()`",
        "operationId": "cancelSignup",
        "tags": [
          "微信前端/活动报名与审核",
          "活动报名与审核"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activityId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/activity/list": {
      "get": {
        "summary": "获取活动列表",
        "deprecated": false,
        "description": "获取活动列表（按角色、标签页不同）\n代码位置: `pages/activity/activity.js` - `loadActivityList()`",
        "operationId": "getActivityList",
        "tags": [
          "微信前端/活动展示",
          "活动展示"
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "example": "student|parent|admin|counselor|teacher"
            }
          },
          {
            "name": "tabIndex",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/activity/detail": {
      "get": {
        "summary": "获取活动详情",
        "deprecated": false,
        "description": "获取单个活动详情\n代码位置: `pages/activity/detail.js` - `loadDetail()`",
        "operationId": "getActivityDetail",
        "tags": [
          "微信前端/活动展示",
          "活动展示"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "example": "student|parent|admin|counselor|teacher"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/activity/my-list": {
      "get": {
        "summary": "我的活动列表",
        "deprecated": false,
        "description": "获取当前用户已报名或参加的活动\n代码位置: `pages/activity/my-activity.js` - `loadRealData()`",
        "operationId": "myActivities",
        "tags": [
          "微信前端/活动展示",
          "活动展示"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "example": "all|upcoming|ongoing|completed|cancelled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/assessment/export": {
      "post": {
        "summary": "科研数据导出",
        "deprecated": false,
        "description": "导出多量表多维度数据用于科研\n代码位置: `pages/assessment/assessment.js` - `realExport()`",
        "operationId": "exportAssessmentData",
        "tags": [
          "微信前端/科研与导入",
          "科研与导入"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scales": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "dimensions": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "properties": {}
                  },
                  "studentRange": {
                    "type": "string",
                    "example": "all|college|grade|custom"
                  },
                  "college": {
                    "type": "string"
                  },
                  "grade": {
                    "type": "string"
                  },
                  "customIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "includeRaw": {
                    "type": "boolean"
                  },
                  "includeStandard": {
                    "type": "boolean"
                  },
                  "includeDemographics": {
                    "type": "boolean"
                  },
                  "format": {
                    "type": "string",
                    "example": "csv"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件流（CSV/Excel）",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/assessment/start": {
      "post": {
        "summary": "开始测评",
        "deprecated": false,
        "description": "开始一次测评（获取题目）\n代码位置: `pages/assessment/process.js` - `loadDetail()`",
        "operationId": "startAssessment",
        "tags": [
          "微信前端/测评答题",
          "测评答题"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scaleId": {
                    "type": "string"
                  },
                  "studentId": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "example": "preQueue|normal"
                  },
                  "consultantId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/assessment/submit": {
      "post": {
        "summary": "提交测评答案",
        "deprecated": false,
        "description": "提交测评答案并计算结果\n代码位置: `pages/assessment/process.js` - `submit()`",
        "operationId": "submitAssessment",
        "tags": [
          "微信前端/测评答题",
          "测评答题"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assessmentId": {
                    "type": "string"
                  },
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionId": {
                          "type": "string"
                        },
                        "optionValue": {
                          "type": "string"
                        },
                        "score": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/assessment/result": {
      "get": {
        "summary": "获取测评结果",
        "deprecated": false,
        "description": "获取已完成的测评结果\n代码位置: `pages/assessment/result.js` - `loadResult()`",
        "operationId": "getAssessmentResult",
        "tags": [
          "微信前端/测评答题",
          "测评答题"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/assessment/my-list": {
      "get": {
        "summary": "获取我的测评列表",
        "deprecated": false,
        "description": "获取当前用户的测评记录\n代码位置: `pages/assessment/my-assessment.js` - `loadRealData()`",
        "operationId": "myAssessments",
        "tags": [
          "微信前端/测评答题",
          "测评答题"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "description": "all/recent/history/abnormal"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/assessment/plan/create": {
      "post": {
        "summary": "创建测评计划",
        "deprecated": false,
        "description": "管理员创建测评计划\n代码位置: `pages/assessment/assessment.js` - `apiCreatePlan()`",
        "operationId": "createAssessmentPlan",
        "tags": [
          "微信前端/测评计划",
          "测评计划"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "endDate": {
                    "type": "string"
                  },
                  "targetColleges": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "targetGrades": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "scales": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "required": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}